Implantation endpoint (service, controller, dto)

+++ src/main/java/br/com/iftm/adsge/pibackend/model/pk/ModuleImplantationId.java
@@ -18,5 +18,5 @@ public class ModuleImplantationId implements Serializable {
     private Integer userId;
 
     @Column(name = "implantation_id")
-    private Integer implantationId;
+    private Long implantationId;
 }
